+++ src/special_asus.rs
@@ -8,9 +8,8 @@ use std::{
 use tokio::time::sleep;
 
 use crate::{
-    config::create_modprobe_conf,
     error::GfxError,
-    pci_device::{rescan_pci_bus, DiscreetGpu, GfxMode, HotplugState},
+    pci_device::{rescan_pci_bus, GfxMode},
 };
 
 const ASUS_DGPU_DISABLE_PATH: &str = "/sys/devices/platform/asus-nb-wmi/dgpu_disable";
@@ -100,6 +99,13 @@ pub fn get_asus_gpu_mux_mode() -> Result<AsusGpuMuxMode, GfxError> {
     ))
 }
 
+pub fn asus_gpu_mux_set_igpu(igpu_on: bool) -> Result<(), GfxError> {
+    debug!("asus_gpu_mux_set_igpu: {igpu_on}");
+    asus_gpu_toggle(igpu_on, ASUS_GPU_MUX_PATH)?;
+    debug!("asus_gpu_mux_set_igpu: success");
+    Ok(())
+}
+
 pub fn asus_dgpu_exists() -> bool {
     if Path::new(ASUS_DGPU_DISABLE_PATH).exists() {
         return true;
@@ -123,6 +129,7 @@ pub fn asus_dgpu_disabled() -> Result<bool, GfxError> {
 
 /// Special ASUS only feature. On toggle to `off` it will rescan the PCI bus.
 pub fn asus_dgpu_set_disabled(disabled: bool) -> Result<(), GfxError> {
+    debug!("asus_dgpu_set_disabled: {disabled}");
     // There is a sleep here because this function is generally called after a hotplug
     // enable, and the deivces require at least a touch of time to finish powering up/down
     std::thread::sleep(Duration::from_millis(500));
@@ -133,6 +140,7 @@ pub fn asus_dgpu_set_disabled(disabled: bool) -> Result<(), GfxError> {
         std::thread::sleep(Duration::from_millis(50));
         rescan_pci_bus()?;
     }
+    debug!("asus_dgpu_set_disabled: success");
     Ok(())
 }
 
@@ -145,6 +153,7 @@ pub fn asus_egpu_exists() -> bool {
 
 /// Special ASUS only feature. On toggle to `on` it will rescan the PCI bus.
 pub fn asus_egpu_set_enabled(enabled: bool) -> Result<(), GfxError> {
+    debug!("asus_egpu_set_enabled: {enabled}");
     // There is a sleep here because this function is generally called after a hotplug
     // enable, and the deivces require at least a touch of time to finish powering up
     std::thread::sleep(Duration::from_millis(500));
@@ -155,6 +164,7 @@ pub fn asus_egpu_set_enabled(enabled: bool) -> Result<(), GfxError> {
         std::thread::sleep(Duration::from_millis(50));
         rescan_pci_bus()?;
     }
+    debug!("asus_egpu_set_enabled: success");
     Ok(())
 }
 
@@ -173,24 +183,22 @@ fn asus_gpu_toggle(status: bool, path: &str) -> Result<(), GfxError> {
 
 /// To be called in main reload code. Specific actions required for asus laptops depending
 /// on is dgpu_disable or mux is available
-pub async fn asus_reload(
+pub async fn asus_boot_safety_check(
     mode: GfxMode,
     asus_use_dgpu_disable: bool,
-    always_reboot: bool,
-    dgpu: &DiscreetGpu,
-) -> Result<(), GfxError> {
-    debug!("asus_reload: asus_use_dgpu_disable: {asus_use_dgpu_disable}, always_reboot: {always_reboot}");
+) -> Result<GfxMode, GfxError> {
+    debug!("asus_reload: asus_use_dgpu_disable: {asus_use_dgpu_disable}");
     // This is a bit of a crap cycle to ensure that dgpu_disable is there before setting it.
     if asus_use_dgpu_disable && !asus_dgpu_exists() {
         if !create_asus_modules_load_conf()? {
             warn!(
-                "asus_reload: Reboot required due to {} creation",
+                "asus_boot_safety_check: Reboot required due to {} creation",
                 ASUS_MODULES_LOAD_PATH
             );
             // let mut cmd = Command::new("reboot");
             // cmd.spawn()?;
         }
-        warn!("asus_reload: asus_use_dgpu_disable is set but asus-wmi appear not loaded yet. Trying for 3 seconds. If there are issues you may need to add asus_nb_wmi to modules.load.d");
+        warn!("asus_boot_safety_check: HotPlug type Asus is set but asus-wmi appear not loaded yet. Trying for 3 seconds. If there are issues you may need to add asus_nb_wmi to modules.load.d");
         let mut count = 3000 / 50;
         while !asus_dgpu_exists() && count != 0 {
             sleep(Duration::from_millis(50)).await;
@@ -201,21 +209,22 @@ pub async fn asus_reload(
     if has_asus_gpu_mux() {
         if let Ok(mux_mode) = get_asus_gpu_mux_mode() {
             if mux_mode == AsusGpuMuxMode::Discreet {
-                create_modprobe_conf(GfxMode::Hybrid, dgpu)?;
-
-                info!("asus_reload: ASUS GPU MUX is in discreet mode");
+                info!("asus_boot_safety_check: ASUS GPU MUX is in discreet mode");
                 if asus_dgpu_exists() {
                     if let Ok(d) = asus_dgpu_disabled() {
                         if d {
-                            error!("asus_reload: dgpu_disable is on while gpu_mux_mode is descrete, can't continue safely, attempting to set dgpu_disable off");
+                            error!("asus_boot_safety_check: dgpu_disable is on while gpu_mux_mode is descrete, can't continue safely, attempting to set dgpu_disable off");
                             asus_dgpu_set_disabled(false)?;
-                            panic!("asus_reload: dgpu_disable is on while gpu_mux_mode is descrete, can't continue safely. Check logs");
+                            panic!("asus_boot_safety_check: dgpu_disable is on while gpu_mux_mode is descrete, can't continue safely. Check logs");
                         } else {
-                            info!("asus_reload: dgpu_disable is off");
+                            info!("asus_boot_safety_check: dgpu_disable is off");
                         }
                     }
                 }
-                return Ok(());
+                return Ok(GfxMode::AsusMuxDgpu);
+            } else if mux_mode == AsusGpuMuxMode::Optimus && mode == GfxMode::AsusMuxDgpu {
+                warn!("asus_boot_safety_check: MUX is in Optimus mode but mode is set to AsusMuxDgpu. Switching to Hybrid");
+                return Ok(GfxMode::Hybrid);
             }
         }
     }
@@ -224,64 +233,12 @@ pub async fn asus_reload(
     if asus_dgpu_exists() {
         // If dgpu_disable is hard set then users won't have a dgpu at all, try set dgpu enabled
         if !asus_use_dgpu_disable && asus_dgpu_disabled()? && mode == GfxMode::Hybrid {
-            warn!("It appears dgpu_disable is true on boot with !asus_use_dgpu_disable, will attempt to re-enable dgpu");
+            warn!("It appears dgpu_disable is true on boot with HotPlug type not set to Asus, will attempt to re-enable dgpu");
             asus_dgpu_set_disabled(false)
                 .map_err(|e| error!("asus_dgpu_set_disabled: {e:?}"))
                 .ok();
         }
     }
 
-    // if asus_use_dgpu_disable && !always_reboot && asus_dgpu_exists() {
-    //     warn!("Has ASUS dGPU and dgpu_disable, toggling hotplug power off/on to prep");
-    //     if dgpu.vendor() == crate::pci_device::GfxVendor::Nvidia {
-    //         crate::kill_nvidia_lsof()?;
-    //         dgpu.do_driver_action("rmmod")?;
-    //     } else if dgpu.vendor() == crate::pci_device::GfxVendor::Amd {
-    //         dgpu.unbind_remove()?;
-    //     }
-    //     dgpu.set_hotplug(HotplugState::Off)?;
-    //     dgpu.set_hotplug(HotplugState::On)?;
-    // }
-    Ok(())
-}
-
-/// To be called in main mode set code. Set some asus specific items depending on mode or
-/// if asus switches are enabled.
-///
-/// Device must be either unbound or drivers unloaded before calling this
-pub fn asus_set_mode(
-    mode: GfxMode,
-    asus_use_dgpu_disable: bool,
-    devices: &mut DiscreetGpu,
-) -> Result<(), GfxError> {
-    debug!("asus_set_mode: {mode:?}, asus_use_dgpu_disable: {asus_use_dgpu_disable}");
-    match mode {
-        GfxMode::Hybrid => {
-            devices.set_hotplug(HotplugState::On)?;
-            if asus_dgpu_exists() && asus_use_dgpu_disable {
-                asus_dgpu_set_disabled(false)?;
-            }
-            if asus_egpu_exists() {
-                asus_egpu_set_enabled(false)?;
-            }
-        }
-        GfxMode::Integrated => {
-            devices.set_hotplug(HotplugState::Off)?;
-            // This can only be done *after* the drivers are removed or a
-            // hardlock will be caused
-            if asus_dgpu_exists() && asus_use_dgpu_disable {
-                asus_dgpu_set_disabled(true)?;
-            }
-            if asus_egpu_exists() {
-                asus_egpu_set_enabled(false)?;
-            }
-        }
-
-        GfxMode::Egpu => {
-            devices.set_hotplug(HotplugState::Off)?;
-            asus_egpu_set_enabled(true)?;
-        }
-        GfxMode::AsusMuxDiscreet | GfxMode::Vfio | GfxMode::None => {}
-    }
-    Ok(())
+    Ok(mode)
 }